move useClinicPatientsFilters into own file
